+++ lib/commands/server_config.mli
@@ -41,6 +41,7 @@ type imapConfig = {
   user_cert_path : string; (* user's certificate/key location *)
   log : string; (* log location, default /var/log *)
   log_level:[`Error|`Info1|`Info2|`Info3|`Debug]; (* log level, default error *)
+  idle_interval: float; (* wait (sec) between idle 'still here' notifications, default 120 sec *)
 }
 
 val default_config : imapConfig